+++ src/solver.cpp
@@ -28,7 +28,7 @@ void Solver::calculate_cost_matrix()
     }
 }
 
-float Solver::get_cost()
+double Solver::get_cost()
 {
     return cost;
 }
@@ -42,14 +42,14 @@ void Solver::solve(int start)
     printf("%f\n", cost);
 }
 
-float Solver::tsp(int pos)
+double Solver::tsp(int pos)
 {
     visited[pos] = order;
     order++;
     if (order == 100) {
         return cost_matrix[pos][start_node];
     }
-    float res = MY_MAX_FLOAT;
+    double res = MY_MAX_FLOAT;
     for (int i = 0; i < NUMBER_OF_POINTS; i++)
     {
         if ((visited[i] < 0) && (cost_matrix[i][pos] > 0))
